requirements.txt
google-cloud-bigquery
pandas-gbq
altair